Have visualization and fanart simultaneously
Now you can have visualization and fanart simultaneous like in the original skin.
The visualization was "blocking" the fanart so placed the visualization first. 
But then the background color of the fanart was blocking the visualization, so I used the "Background_Opacity" to fix that.

Possible side-effect: the "Background_Opacity" also has influence even if visualization is disabled.