This PR adds a button on the DialogPVRInfo screen to utilise and call the KN Switchtimer e.g to be able to set reminders and automatically tune to a PVR channel.
Enabling the KN Switchtimer does introduce a new value in the PVR context menu, However this isn't particularly user friendly so i believe the most natural place to add a button to call this service is along with the switch, record, add timer buttons on the DialogPVRInfo screen.

if KN Switchtimer addon  is not installed then the button is hidden.

still to do:
 - Find a more appropriate image for button
 - Add possible entry to skin addon support screen